requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.1.2
django-leaflet==0.27.1
django-templated-mail==1.1.1
django-webpack-loader==0.7.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
gunicorn==20.0.4
numpy==1.19.2
oauthlib==3.1.1
pandas==1.1.3
postgis==1.0.4
psycopg2==2.8.6
python-dateutil==2.8.1
requests==2.26.0
requests-oauthlib==1.3.0
pytz==2020.1
six==1.14.0
SQLAlchemy==1.3.20
sqlparse==0.4.1
typing-extensions==3.10.0.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
whitenoise==5.2.0
cffi==1.14.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
defusedxml==0.7.1
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
pycparser==2.20
uritemplate==3.0.1
zipp==3.5.0
PyJWT==2.1.0
python3-openid==3.2.0
